fixing major bottleneck in adf.apply
using eval(call(lm, ...)) was causing the entire dataset
to be copied multiple times. I fixed this by instead dealing
with 'subset' and 'na.action' manually and creating four
explicit cases for all combinations of weights & offset
being supplied. The run time is now approximatley 5x faster
this was truly a major performance bottleneck. I've updated
the version number of 0.3 so that we can make sure that any
simulations are using the updated version.